рефактор ком-та Friends
